+++ CHANGELOG.md
@@ -6,7 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Changed
+- Improve responsiveness of viewpoint switching, and retain playing state.
+
 ### Added
+
 ### Fixed
 
 ## [6.2.1] - 2024-12-08